Delete ReservationsController and update routes.rb
